%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  getArchiveRootPath.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setArchiveRootPath}
\alias{setArchiveRootPath.default}
\alias{setArchiveRootPath}

\title{Sets the root path to the file archive directory}

\description{
 Sets the root path to the file archive directory.
}

\usage{\method{setArchiveRootPath}{default}(path="~/.Rarchive", ...)}

\arguments{
  \item{path}{The path.}
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) the old root path.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
 \code{\link{getArchiveRootPath}}().
}



\keyword{programming}
\keyword{IO}
